+++ src/main/java/nextstep/courses/domain/session/enrollment/EnrollmentInfo.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.session.enrollment2;
+package nextstep.courses.domain.session.enrollment;
 
 import nextstep.courses.domain.session.Student;
 import nextstep.courses.domain.session.Students;
