+++ src/main/java/nextstep/courses/infrastructure/JdbcSessionsRepository.java
@@ -4,7 +4,7 @@ import nextstep.courses.domain.session.Session;
 import nextstep.courses.domain.session.SessionInfo;
 import nextstep.courses.domain.session.Sessions;
 import nextstep.courses.domain.session.SessionsRepository;
-import nextstep.courses.domain.session.enrollment2.EnrollmentInfo;
+import nextstep.courses.domain.session.enrollment.EnrollmentInfo;
 import org.springframework.jdbc.core.JdbcOperations;
 import org.springframework.stereotype.Repository;
